6-size.c

+++ 0x00-hello_world/6-size.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 /**
  * main - program to print size of various computer types
- * Return Always 0 (Success)
+ * Return: 0 
  */
 int main(void)
 {
@@ -9,6 +9,6 @@ printf("size of a char: %i byte(s)\n", sizeof(char));
 printf("size of an int: %i byte(s)\n", sizeof(int));
 printf("size of a long int: %i byte(s)\n", sizeof(long int));
 printf("size of a long long int: %i byte(s)\n", sizeof(long long int));
-printf("size of a float: %i byte(s)\n", (unsigned long)sizeof(float));
+printf("size of a float: %i byte(s)\n", sizeof(float));
 return (0);
 }
